memorycard: drop redundant s16 cast on signed m_artifacts in MakeSaveData (83.76->83.82)

+++ src/memorycard.cpp
@@ -1702,7 +1702,7 @@ void CMemoryCardMan::MakeSaveData()
         {
             const int slot = artifact >> 5;
             const u32 bit = 1u << (artifact & 31);
-            if (static_cast<s16>(caravanWork->m_artifacts[artifact]) > 0)
+            if (caravanWork->m_artifacts[artifact] > 0)
             {
                 *reinterpret_cast<u32*>(dst + 0xBC + slot * 4) |= bit;
             }
@@ -1710,7 +1710,7 @@ void CMemoryCardMan::MakeSaveData()
             const int artifact1 = artifact + 1;
             const int slot1 = artifact1 >> 5;
             const u32 bit1 = 1u << (artifact1 & 31);
-            if (static_cast<s16>(caravanWork->m_artifacts[artifact1]) > 0)
+            if (caravanWork->m_artifacts[artifact1] > 0)
             {
                 *reinterpret_cast<u32*>(dst + 0xBC + slot1 * 4) |= bit1;
             }
@@ -1718,7 +1718,7 @@ void CMemoryCardMan::MakeSaveData()
             const int artifact2 = artifact + 2;
             const int slot2 = artifact2 >> 5;
             const u32 bit2 = 1u << (artifact2 & 31);
-            if (static_cast<s16>(caravanWork->m_artifacts[artifact2]) > 0)
+            if (caravanWork->m_artifacts[artifact2] > 0)
             {
                 *reinterpret_cast<u32*>(dst + 0xBC + slot2 * 4) |= bit2;
             }
